Add note about PascalCasing BEM blocks with React
BEM wasn't designed with React in mind, and I think that this deviation
from the convention is useful to more closely connect the concepts of
BEM to your React components. Granted, with CSS Modules or inline CSS
(CSS in JS), this is pretty irrelevant, but it is nice to provide the
guidance for folks who are not yet living in that world.