Include a sass-lint configuration file
This sass-lint configuration file follows the documented styleguide as
best we can with the current sass-lint rules available, and makes some
notes as to other items that could, in principle, be added as automated
rules.

Note that http://sasstools.github.io/make-sass-lint-config/ was used as
the starting point of this, but manual revision was necessary to get a
good result.

To help developers who are retrospectively applying this styleguide,
the rules that would require non-trivial changes (use of id selectors,
class name formats, nesting levels, and variable names) are flagged as
warnings rather than errors.